changes the column type for discord id to accommodate discord id size

+++ db/migrate/20170718134543_create_players.rb
@@ -1,7 +1,7 @@
 class CreatePlayers < ActiveRecord::Migration[5.1]
   def change
     create_table :players do |t|
-      t.integer :discord_id, null: false
+      t.bigint :discord_id, null: false
       t.string :username, null: false
       t.integer :elo, null: false, default: 1200
       t.timestamps
